backend/requirements.txt
 Flask==3.0.0
 flask-cors==4.0.0
 pymongo==4.6.0
 python-dotenv==1.0.0
 ffmpeg-python==0.2.0
 bson==0.5.10